Move WalletDAOs to src
