+++ wallet/src/main/scala/org/bitcoins/wallet/models/WalletDAOs.scala
@@ -1,0 +1,26 @@
+package org.bitcoins.wallet.models
+
+case class WalletDAOs(
+    accountDAO: AccountDAO,
+    addressDAO: AddressDAO,
+    addressTagDAO: AddressTagDAO,
+    utxoDAO: SpendingInfoDAO,
+    transactionDAO: TransactionDAO,
+    incomingTxDAO: IncomingTransactionDAO,
+    outgoingTxDAO: OutgoingTransactionDAO,
+    scriptPubKeyDAO: ScriptPubKeyDAO,
+    stateDescriptorDAO: WalletStateDescriptorDAO
+) {
+
+  val list = Vector(
+    scriptPubKeyDAO,
+    accountDAO,
+    addressDAO,
+    addressTagDAO,
+    transactionDAO,
+    incomingTxDAO,
+    utxoDAO,
+    outgoingTxDAO,
+    stateDescriptorDAO
+  )
+}
